+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.3
+- [enhancement] Load GMate welcome page upon initial installation
+
 ## 0.4.2
 - [enhancement] Automatically retire tasks for deleted events
 - [enhancement] Allow user to cancel the current task from the "room-found" notification